+++ wordpress/wp-content/plugins/font-awesome-4-menus/n9m-font-awesome-4.php
@@ -3,7 +3,7 @@
 Plugin Name: Font Awesome 4 Menus
 Plugin URI: http://www.newnine.com/plugins/font-awesome-4-menus
 Description: Join the retina/responsive revolution by easily adding Font Awesome 4.3 icons to your WordPress menus and anywhere else on your site! No programming necessary.
-Version: 4.3.0.1
+Version: 4.3.0.3
 Author: New Nine Media
 Author URI: http://www.newnine.com
 License: GPLv2 or later